ml_pipelines/ema_pipeline/requirements.txt
alpaca-py
python-dotenv
pandas  
requests
